Refactor Ecobee.__init__
Remove mandatory sync calls to request_pin() and update(); this is a breaking change for users of this library that expect for a pin to be obtained or self.thermostats to be populated automatically when the object is created. Instead, the caller will need to request a pin and then update manually as part of its own flow.