Add token_refresh logic to request calls
This change adds a wrapper call around _request for use
with calls that rely on a valid token.
If the request fails due to an expired token, the refresh_tokens
call is made and the _request is tried again. If it still fails
pass the state up to the caller.
This also updates the request library 2.25 or greater